I've finished the states, now the actions remain. Left and right work; I should work on walk and take next. Then, I'll work on multiplayer functionality and on player interaction actions such as talk and shout.
Documentation for adding actions:

Where should I initialize this action? Maybe I can just give it the dungeon in the constructor? Or should I create a factory for it?

I decided to implement the walking logic in the Dungeon class. Now I'm thinking about how to return the reply to Act. How did I do it before?

// TODO: Unregister and reregister the player to the next room
Should editing the player members be done in the Dungeon class? I think not!

I remember there is a nice way to do assertions that throw a text message. This should be changed:

const Number Wall::GetNextRoomNumber() const {
    assert(IsDoor());
    if (!IsDoor()) {
        throw std::logic_error("GetNextRoomNumber() called on a non-door wall");
    }
    return m_nextRoomNumber;
} // It doesn't make sense to throw something after an assertion
Bugs: It seems that walk works and the player does go to a different room, but:

The drawing of the room always shows room 0.
It says, "You've entered room," without specifying the number.